@@ -10,37 +10,6 @@
             "email": "norbert@orzechowicz.pl"
         }
     ],
-    "require": {
-        "php": "^7.1|^8.0",
-        "doctrine/dbal": "^2.6",
-        "doctrine/doctrine-bundle": "^1.4|^2.0",
-        "symfony/dependency-injection": "^3.4|^4.0|^5.0",
-        "symfony/doctrine-bridge": "^3.4|^4.0|^5.0",
-        "symfony/form" : "^3.4|^4.0|^5.0",
-        "symfony/framework-bundle": "^3.4|^4.0|^5.0",
-        "symfony/property-access": "^3.4|^4.0|^5.0",
-        "symfony/validator": "^3.4|^4.0|^5.0",
-        "twig/twig": "^1.34|^2.0"
-    },
-    "require-dev": {
-        "doctrine/orm": "^2.5",
-        "friendsofsymfony/ckeditor-bundle": "^1.0|^2.0",
-        "fsi/doctrine-extensions-bundle": "^1.0|^2.0",
-        "fsi/form-extensions-bundle": "^2.0",
-        "phpunit/phpunit": "^7.1|^8.0",
-        "phpspec/phpspec": "^5.0|^6.0",
-        "phpstan/phpstan": "^0.12.66",
-        "phpstan/phpstan-phpunit": "^0.12.15",
-        "squizlabs/php_codesniffer": "^3.5",
-        "symfony/phpunit-bridge": "^4.3|^5.0"
-    },
-    "suggest": {
-        "fsi/doctrine-extensions-bundle": "For using fsi_file resource type",
-        "friendsofsymfony/ckeditor-bundle": "For using ckeditor resource type"
-    },
-    "config": {
-        "bin-dir": "vendor/bin"
-    },
     "autoload": {
         "psr-4": { "FSi\\Bundle\\ResourceRepositoryBundle\\": "" }
     },
@@ -49,6 +18,39 @@
             "spec\\": "spec",
             "FSi\\Bundle\\ResourceRepositoryBundle\\Tests\\": "tests"
         }
+    },
+    "require": {
+        "php": "^7.4|^8.0",
+        "doctrine/dbal": "^2.8",
+        "doctrine/doctrine-bundle": "^1.12|^2.3",
+        "symfony/dependency-injection": "^4.4|^5.0",
+        "symfony/doctrine-bridge": "^4.4|^5.0",
+        "symfony/form" : "^4.4|^5.0",
+        "symfony/framework-bundle": "^4.4|^5.0",
+        "symfony/property-access": "^4.4|^5.0",
+        "symfony/validator": "^4.4|^5.0",
+        "twig/twig": "^2.13"
+    },
+    "require-dev": {
+        "doctrine/orm": "^2.5",
+        "friendsofsymfony/ckeditor-bundle": "^2.1",
+        "fsi/doctrine-extensions-bundle": "^2.1",
+        "fsi/files": "^1.1",
+        "fsi/form-extensions-bundle": "^2.0",
+        "phpunit/phpunit": "^8.0",
+        "phpspec/phpspec": "^6.0",
+        "phpstan/phpstan": "^0.12.66",
+        "phpstan/phpstan-phpunit": "^0.12.15",
+        "squizlabs/php_codesniffer": "^3.5",
+        "symfony/phpunit-bridge": "^5.2"
+    },
+    "suggest": {
+        "fsi/doctrine-extensions-bundle": "For using fsi_file resource type",
+        "fsi/files": "For using web_file resource type",
+        "friendsofsymfony/ckeditor-bundle": "For using ckeditor resource type"
+    },
+    "config": {
+        "bin-dir": "vendor/bin"
     },
     "extra": {
         "branch-alias": {
